Update sort012.java
previously this was using Dutch National Flag Algorithm OR 3-way Partitioning. but which can be hard to understand for beginners so here we are just counting numbers of 0's, 1's, and 2's. and then just assigning the number of respective 0,1 and 2's in the array. asymptotically both will run at the same time but this approach is easy to understand and learn.